requirements.txt
-i https://pypi.org/simple
attrs==21.4.0
beautifulsoup4==4.11.1
cffi==1.15.0
click==8.1.3
colorama==0.4.5
coverage==6.4.1
execnet==1.9.0
flask==2.1.2
future==0.18.2
gunicorn==20.1.0
iniconfig==1.1.1
itsdangerous==2.1.2
jinja2==3.1.2
lxml==4.9.0
mando==0.6.4
markupsafe==2.1.1
numpy==1.22.4
packaging==21.3
pandas==1.4.2
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pynacl==1.5.0
pyparsing==3.0.9
pytest==7.1.2
pytest-forked==1.4.0
pytest-xdist==2.5.0
python-dateutil==2.8.2
pytz==2022.1
radon==5.1.0
setuptools==62.5.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tomli==2.0.1
werkzeug==2.1.2
